Move contents of Internal folder to Mapping folder
These classes are all related to mapping for the EnumerableDataReader
