+++ src/BulkWriter.Tests/TypeExtensionsTests.cs
@@ -1,7 +1,7 @@
 ﻿using System.ComponentModel.DataAnnotations;
 using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
-using BulkWriter.Internal;
+using BulkWriter.Mapping;
 using Xunit;
 
 namespace BulkWriter.Tests
