+++ src/BulkWriter/Mapping/GetPropertyValueHandler.cs
@@ -1,4 +1,4 @@
-﻿namespace BulkWriter.Internal
+﻿namespace BulkWriter.Mapping
 {
     internal delegate object GetPropertyValueHandler(object instance);
 }
