+++ src/BulkWriter/Mapping/MappingDestination.cs
@@ -1,4 +1,4 @@
-﻿namespace BulkWriter.Internal
+﻿namespace BulkWriter.Mapping
 {
     internal class MappingDestination
     {
