+++ src/BulkWriter/Mapping/PropertyMapping.cs
@@ -1,4 +1,4 @@
-﻿namespace BulkWriter.Internal
+﻿namespace BulkWriter.Mapping
 {
     internal class PropertyMapping
     {
